Simplify device interface
