+++ src/Form/listenForFormSubmission.js
@@ -5,6 +5,11 @@ const listenForGlobalFormSubmission = () => {
     if (!isApp) return
 
     try {
+        window.addEventListener('submit', (e) =>
+            // @ts-ignore
+            forms.get(e.target)?.submitHandler(),
+        true)
+
         const observer = new PerformanceObserver((list) => {
             const entries = list.getEntries().filter((entry) =>
                 // @ts-ignore why does TS not know about `entry.initiatorType`?
