+++ src/autofill.js
@@ -1,18 +1,18 @@
+// Polyfills/shims
+require('./requestIdleCallback');
+
 (() => {
     try {
         if (!window.isSecureContext) return
+        const deviceInterface = require('./DeviceInterface')
 
-        const listenForGlobalFormSubmission = require('./Form/listenForFormSubmission')
         const {processConfig} = require('@duckduckgo/content-scope-scripts/src/apple-utils')
         const {autofillEnabled} = require('./autofill-utils')
-
-        const inject = require('./inject')
 
         // chrome is only present in desktop browsers
         if (typeof chrome === 'undefined') {
             if (autofillEnabled(processConfig)) {
-                listenForGlobalFormSubmission()
-                inject()
+                deviceInterface.init()
             }
         } else {
             // Check if the site is marked to skip autofill
@@ -23,7 +23,7 @@
                 },
                 (response) => {
                     if (!response?.site?.brokenFeatures?.includes('autofill')) {
-                        inject()
+                        deviceInterface.init()
                     }
                 }
             )
